S&P500 data now supports "historical" option
   >> get_sp500_symbols(historical=True)

This will retrieves a more complete list of companies contained in the
S&P 500 from Wikipedia.  It parses the "Recent changes to the list
of S&P 500 Components" and adds a start/end date to symbols, as applicable.

The historical flag also adds three new keys to each company's object:
    start: A None value here means "since the beginning, as far as we know"
           Otherwise, the date on which the company was added
    end: None here means the company is still a part of the S&P500.
         Otherwise, the date on which the company was removed.
    current:  If the company is currently listed on the S&P 500.
              Probably redundant with end=None.